Convert uploadFiles logic into Wizard + general refactor/re-organize
