+++ src/commands/uploadFiles/IExistingFileContext.ts
@@ -1,0 +1,6 @@
+import { IActionContext } from "@microsoft/vscode-azext-utils";
+
+export interface IExistingFileContext extends IActionContext {
+    localFilePath: string;
+    remoteFilePath: string;
+}
